requirements.txt
alembic==1.13.1
aniso8601==9.0.1
APScheduler==3.10.4
attrs==23.2.0
backports.zoneinfo==0.2.1
blinker==1.8.2
click==8.1.7
exceptiongroup==1.2.1
flask==3.0.3
Flask-APScheduler==1.13.1
Flask-JWT-Extended==4.6.0
flask-mail==0.10.0
Flask-Migrate==4.0.7
flask-restx==1.3.0
flask-sqlalchemy==3.1.1
greenlet==3.0.3
importlib-metadata==7.1.0
importlib-resources==6.4.0
iniconfig==2.0.0
itsdangerous==2.2.0
jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
Mako==1.3.5
MarkupSafe==2.1.5
packaging==24.0
pkgutil-resolve-name==1.3.10
pluggy==1.5.0
PyJWT==2.8.0
pytest==8.2.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
referencing==0.35.1
rpds-py==0.18.1
six==1.16.0
SQLAlchemy==2.0.30
tomli==2.0.1
typing-extensions==4.11.0
tzlocal==5.2
werkzeug==3.0.3
zipp==3.18.2
